oikeusryhtmat

+++ protected/controllers/OikeusRyhmatController.php
@@ -152,7 +152,7 @@ class OikeusRyhmatController extends Controller
 	public function actionDelete($id)
 	{
 		$model=$this->loadModel($id);
-		//if($model->for_delete == 'true')
+		if($model->for_delete == 'true')
 			$model->delete();
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
